Endpoint to list all users
Implement endpoint to list all users:
- Endpoint GET `/api/users`
- Supports pagination, searching by name and ordering `asc` or `desc`
- Displays the `salary` that is active for the users that have one
- API Tests to prove it's working

Note: For the users that don't have a salary, for now it's shown as `null`, but the idea is
to show the most recent inactive salary. I'm still thinking on solutions, but to avoid
being stuck, I'll proceed implementing the rest of the challenge.