Endpoint to invite users and check invite job status
Implemented two endpoints:
- One to invite users. To avoid having the client waiting for a response, the
task is run async, a job id is returned to the client so he can check how's
the job gjoing.
- One to check the status of a job. The ID the client has received previously
should be passed, and it will be returned some basic info about the progress
of the users invitation.

 # Please enter the commit message for your changes. Lines
starting # with '#' will be ignored, and an empty message aborts the commit.